Semaine1/hugging face app/requirements.txt
gradio
Pillow
numpy
opencv-python
matplotlib
scikit-image
scipy
